Start reorg of per_epoch_processing
