+++ eth2/state_processing/src/per_block_processing/errors.rs
