+++ eth2/state_processing/src/per_block_processing/verify_proposer_slashing.rs
@@ -1,4 +1,4 @@
-use crate::errors::{ProposerSlashingInvalid as Invalid, ProposerSlashingValidationError as Error};
+use super::errors::{ProposerSlashingInvalid as Invalid, ProposerSlashingValidationError as Error};
 use ssz::SignedRoot;
 use types::*;
