+++ eth2/state_processing/src/per_block_processing/verify_slashable_attestation.rs
@@ -1,4 +1,4 @@
-use crate::errors::{
+use super::errors::{
     SlashableAttestationInvalid as Invalid, SlashableAttestationValidationError as Error,
 };
 use ssz::TreeHash;
